factor pyramid_reproject into its own module
